Modified seeds 25d8a9c8 and 25ff71a9
Changed 25d8a9c8 to be more in line with previous examples--changing from `for i, row in enumerate(input_grid.T)` to `for y in range(m)` (where `m` is from `n, m = input_grid.shape`). Changed other lines to adhere to this standard, such as from `for j in range(len(row))` to `for x in range(n)`.

For seed 25ff71a9, removed hardcoded output grid size and `blit` usage to simply a `translate` of the object.