Added a function to common.py to return a new color
Added a function to `common.py` that returns a color not in the list of `not_allowed_colors`, by default does not return `Color.BLACK`.

I found that I would often want to randomly choose a color in `generate_input` functions that isn't already used, with ever-increasing numbers of colors to avoid. This helps that by choosing a color not used.