Added padding functionality to `random_free_location_for_object`
Added a new functionality to check for free locations with padding for an object.
I.e., with padding=1, there would be no contact (rather than no collision with padding=0) between the grid and the sprite.

Works by expanding the given sprite by 1 layer for each level of padding, subject to 4 or 8-way connectivity (i.e., the padding only expands 4 or 8-way on the sprite).
E.g., a 1x1 pixel would turn into a plus sign with one layer of padding and 4-way connectivity, or it would turn into a 3x3 grid with 8-way connectivity.