Added input generation to seed 264363fd
Completed the `input_generation` function for this seed. Uses the new `padding` functionality for `random_free_location_for_object` (see line 221 and 231).

I assumed this would be run with a timeout function. This could potentially turn into an infinite loop if unlucky consistently, but that isn't likely (e.g., if two maximum size rectangles are generated, there would be no way to place both of them in the grid. This would trigger an error in `random_free_location_for_object`, which would be caught, and then the rectangles (and their sizes) would be generated randomly again).