Fixing padding implementation
Padding had issues with the pad around the sprite being correctly applied.
Also fixed `possible_locations` to take padding into account. This combination of `border_size` and `padding` means that if `border_size=1` and `padding=1`, then the sprite will always have at least 2 pixels between it and the border. Previously, padded objects could be on the border as the value of `padding` did not affect `possible_locations`.