Fixed bug in seed `264363fd.py`
Bug occurs in the `main` function when the `rectangle` object is so large that it has more pixels than the grid. Fixed by cropping the rectangle beforehand.
Additionally, modified `input_generation` to require a gap of at least `2` pixels on either axis between special pixels (from `1` previously). This is because the input examples never consider what happens when crosshair patterns overlap, so this removes that consideration from possible inputs (so as to avoid having to make ambiguous choices).

Also re-added border padding to the generated rectangles. In my previous version of `random_free_location_for_object`, `padding` additively applied to the border size when choosing locations. This new version does not do that, thus I need to add a border.